+++ components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineSessionTest.kt
@@ -51,6 +51,7 @@ import mozilla.components.support.test.expectException
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.whenever
 import mozilla.components.support.utils.DownloadUtils.RESPONSE_CODE_SUCCESS
+import mozilla.components.support.utils.FakeDownloadFileUtils
 import mozilla.components.support.utils.ThreadUtils
 import mozilla.components.test.ReflectionUtils
 import org.json.JSONObject
@@ -337,6 +338,7 @@ class GeckoEngineSessionTest {
             mock(),
             geckoSessionProvider = geckoSessionProvider,
             privateMode = true,
+            downloadFileUtils = FakeDownloadFileUtils(guessFileName = { _, _, _ -> "image name.png" }),
         )
 
         val observer: EngineSession.Observer = mock()
@@ -376,6 +378,7 @@ class GeckoEngineSessionTest {
             mock(),
             geckoSessionProvider = geckoSessionProvider,
             privateMode = true,
+            downloadFileUtils = FakeDownloadFileUtils(guessFileName = { _, _, _ -> "image image.png" }),
         )
 
         val observer: EngineSession.Observer = mock()
@@ -416,6 +419,7 @@ class GeckoEngineSessionTest {
             mock(),
             geckoSessionProvider = geckoSessionProvider,
             privateMode = true,
+            downloadFileUtils = FakeDownloadFileUtils(guessFileName = { _, _, _ -> "image.png" }),
         )
 
         val observer: EngineSession.Observer = mock()
@@ -453,6 +457,7 @@ class GeckoEngineSessionTest {
             mock(),
             geckoSessionProvider = geckoSessionProvider,
             privateMode = true,
+            downloadFileUtils = FakeDownloadFileUtils(guessFileName = { _, _, _ -> "image.png" }),
         )
 
         val observer: EngineSession.Observer = mock()
@@ -4619,6 +4624,7 @@ class GeckoEngineSessionTest {
         val engineSession = GeckoEngineSession(
             runtime = mock(),
             geckoSessionProvider = geckoSessionProvider,
+            downloadFileUtils = FakeDownloadFileUtils(guessFileName = { _, _, _ -> "Mozilla.pdf" }),
         ).apply {
             currentUrl = "https://mozilla.org"
             currentTitle = "Mozilla"