+++ components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/util/SpeculativeSessionFactoryTest.kt
@@ -7,6 +7,7 @@ package mozilla.components.browser.engine.gecko.util
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.whenever
+import mozilla.components.support.utils.FakeDownloadFileUtils
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNotSame
@@ -34,11 +35,23 @@ class SpeculativeSessionFactoryTest {
         val factory = SpeculativeSessionFactory()
         assertNull(factory.speculativeEngineSession)
 
-        factory.create(runtime = runtime, private = true, contextId = null, defaultSettings = mock())
+        factory.create(
+            runtime = runtime,
+            private = true,
+            contextId = null,
+            defaultSettings = mock(),
+            downloadFileUtils = FakeDownloadFileUtils(),
+        )
         val speculativeSession = factory.speculativeEngineSession
         assertNotNull(speculativeSession)
 
-        factory.create(runtime = runtime, private = true, contextId = null, defaultSettings = mock())
+        factory.create(
+            runtime = runtime,
+            private = true,
+            contextId = null,
+            defaultSettings = mock(),
+            downloadFileUtils = FakeDownloadFileUtils(),
+            )
         assertSame(speculativeSession, factory.speculativeEngineSession)
     }
 
@@ -47,11 +60,23 @@ class SpeculativeSessionFactoryTest {
         val factory = SpeculativeSessionFactory()
         assertNull(factory.speculativeEngineSession)
 
-        factory.create(runtime = runtime, private = true, contextId = null, defaultSettings = mock())
+        factory.create(
+            runtime = runtime,
+            private = true,
+            contextId = null,
+            defaultSettings = mock(),
+            downloadFileUtils = FakeDownloadFileUtils(),
+            )
         val speculativeSession = factory.speculativeEngineSession
         assertNotNull(speculativeSession)
 
-        factory.create(runtime = runtime, private = false, contextId = null, defaultSettings = mock())
+        factory.create(
+            runtime = runtime,
+            private = false,
+            contextId = null,
+            defaultSettings = mock(),
+            downloadFileUtils = FakeDownloadFileUtils(),
+        )
         assertNotSame(speculativeSession, factory.speculativeEngineSession)
         assertFalse(speculativeSession!!.engineSession.geckoSession.isOpen)
         assertFalse(speculativeSession.engineSession.isObserved())
@@ -62,7 +87,13 @@ class SpeculativeSessionFactoryTest {
         val factory = SpeculativeSessionFactory()
         assertFalse(factory.hasSpeculativeSession())
 
-        factory.create(runtime = runtime, private = true, contextId = null, defaultSettings = mock())
+        factory.create(
+            runtime = runtime,
+            private = true,
+            contextId = null,
+            defaultSettings = mock(),
+            downloadFileUtils = FakeDownloadFileUtils(),
+            )
         assertTrue(factory.hasSpeculativeSession())
 
         val speculativeSession = factory.get(private = true, contextId = null)
@@ -78,7 +109,13 @@ class SpeculativeSessionFactoryTest {
         val factory = SpeculativeSessionFactory()
         assertNull(factory.speculativeEngineSession)
 
-        factory.create(runtime = runtime, private = true, contextId = null, defaultSettings = mock())
+        factory.create(
+            runtime = runtime,
+            private = true,
+            contextId = null,
+            defaultSettings = mock(),
+            downloadFileUtils = FakeDownloadFileUtils(),
+            )
         val speculativeSession = factory.speculativeEngineSession
         assertNotNull(speculativeSession)
 
@@ -90,7 +127,13 @@ class SpeculativeSessionFactoryTest {
     @Test
     fun `clears speculative session on crash`() {
         val factory = SpeculativeSessionFactory()
-        factory.create(runtime = runtime, private = true, contextId = null, defaultSettings = mock())
+        factory.create(
+            runtime = runtime,
+            private = true,
+            contextId = null,
+            defaultSettings = mock(),
+            downloadFileUtils = FakeDownloadFileUtils(),
+            )
         assertTrue(factory.hasSpeculativeSession())
         val speculativeSession = factory.speculativeEngineSession
 
@@ -103,7 +146,13 @@ class SpeculativeSessionFactoryTest {
     @Test
     fun `clears speculative session when process is killed`() {
         val factory = SpeculativeSessionFactory()
-        factory.create(runtime = runtime, private = true, contextId = null, defaultSettings = mock())
+        factory.create(
+            runtime = runtime,
+            private = true,
+            contextId = null,
+            defaultSettings = mock(),
+            downloadFileUtils = FakeDownloadFileUtils(),
+            )
         assertTrue(factory.hasSpeculativeSession())
         val speculativeSession = factory.speculativeEngineSession
 
@@ -116,7 +165,13 @@ class SpeculativeSessionFactoryTest {
     @Test
     fun `clear unregisters observer and closes session`() {
         val factory = SpeculativeSessionFactory()
-        factory.create(runtime = runtime, private = true, contextId = null, defaultSettings = mock())
+        factory.create(
+            runtime = runtime,
+            private = true,
+            contextId = null,
+            defaultSettings = mock(),
+            downloadFileUtils = FakeDownloadFileUtils(),
+            )
         assertTrue(factory.hasSpeculativeSession())
         val speculativeSession = factory.speculativeEngineSession
         assertTrue(speculativeSession!!.engineSession.isObserved())