+++ components/feature/downloads/src/test/java/mozilla/components/feature/downloads/manager/AndroidDownloadManagerTest.kt
@@ -18,6 +18,7 @@ import mozilla.components.feature.downloads.AbstractFetchDownloadService.Compani
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.grantPermission
 import mozilla.components.support.test.robolectric.testContext
+import mozilla.components.support.utils.FakeDownloadFileUtils
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
@@ -48,7 +49,11 @@ class AndroidDownloadManagerTest {
             userAgent = "Mozilla/5.0 (Linux; Android 7.1.1) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Focus/8.0 Chrome/69.0.3497.100 Mobile Safari/537.36",
         )
         store = BrowserStore()
-        downloadManager = AndroidDownloadManager(testContext, store)
+        downloadManager = AndroidDownloadManager(
+            applicationContext = testContext,
+            store = store,
+            downloadFileUtils = FakeDownloadFileUtils(),
+            )
     }
 
     @Test(expected = SecurityException::class)
@@ -103,7 +108,13 @@ class AndroidDownloadManagerTest {
 
     @Test
     fun `GIVEN a device that supports scoped storage THEN permissions must not included file access`() {
-        val downloadManager = spy(AndroidDownloadManager(testContext, store))
+        val downloadManager = spy(
+            AndroidDownloadManager(
+            applicationContext = testContext,
+            store = store,
+            downloadFileUtils = FakeDownloadFileUtils(),
+        ),
+        )
 
         doReturn(Build.VERSION_CODES.Q).`when`(downloadManager).getSDKVersion()
         println(downloadManager.permissions.joinToString { it })
@@ -112,7 +123,13 @@ class AndroidDownloadManagerTest {
 
     @Test
     fun `GIVEN a device does not supports scoped storage THEN permissions must be included file access`() {
-        val downloadManager = spy(AndroidDownloadManager(testContext, store))
+        val downloadManager = spy(
+            AndroidDownloadManager(
+            applicationContext = testContext,
+            store = store,
+            downloadFileUtils = FakeDownloadFileUtils(),
+        ),
+        )
 
         doReturn(Build.VERSION_CODES.P).`when`(downloadManager).getSDKVersion()