+++ components/feature/downloads/src/test/java/mozilla/components/feature/downloads/manager/FetchDownloadManagerTest.kt
@@ -30,6 +30,8 @@ import mozilla.components.support.test.any
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.grantPermission
 import mozilla.components.support.test.robolectric.testContext
+import mozilla.components.support.utils.DownloadFileUtils
+import mozilla.components.support.utils.FakeDownloadFileUtils
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
@@ -319,6 +321,7 @@ class FetchDownloadManagerTest {
         override val downloadEstimator: DownloadEstimator = DownloadEstimator(
             FakeDateTimeProvider(),
         )
+        override val downloadFileUtils: DownloadFileUtils = FakeDownloadFileUtils()
         override val packageNameProvider = FakePackageNameProvider("org.mozilla.fenix.test")
     }
 }